% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.swf_operations.R
\name{list_closed_workflow_executions}
\alias{list_closed_workflow_executions}
\title{Returns a list of closed workflow executions in the specified domain that meet the filtering criteria}
\usage{
list_closed_workflow_executions(domain, startTimeFilter = NULL,
  closeTimeFilter = NULL, executionFilter = NULL,
  closeStatusFilter = NULL, typeFilter = NULL, tagFilter = NULL,
  nextPageToken = NULL, maximumPageSize = NULL, reverseOrder = NULL)
}
\arguments{
\item{domain}{The name of the domain that contains the workflow executions to list.}

\item{startTimeFilter}{If specified, the workflow executions are included in the returned results based on whether their start times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their start times.

\code{startTimeFilter} and \code{closeTimeFilter} are mutually exclusive. You must specify one of these in a request but not both.}

\item{closeTimeFilter}{If specified, the workflow executions are included in the returned results based on whether their close times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their close times.

\code{startTimeFilter} and \code{closeTimeFilter} are mutually exclusive. You must specify one of these in a request but not both.}

\item{executionFilter}{If specified, only workflow executions matching the workflow ID specified in the filter are returned.

\code{closeStatusFilter}, \code{executionFilter}, \code{typeFilter} and \code{tagFilter} are mutually exclusive. You can specify at most one of these in a request.}

\item{closeStatusFilter}{If specified, only workflow executions that match this \emph{close status} are listed. For example, if TERMINATED is specified, then only TERMINATED workflow executions are listed.

\code{closeStatusFilter}, \code{executionFilter}, \code{typeFilter} and \code{tagFilter} are mutually exclusive. You can specify at most one of these in a request.}

\item{typeFilter}{If specified, only executions of the type specified in the filter are returned.

\code{closeStatusFilter}, \code{executionFilter}, \code{typeFilter} and \code{tagFilter} are mutually exclusive. You can specify at most one of these in a request.}

\item{tagFilter}{If specified, only executions that have the matching tag are listed.

\code{closeStatusFilter}, \code{executionFilter}, \code{typeFilter} and \code{tagFilter} are mutually exclusive. You can specify at most one of these in a request.}

\item{nextPageToken}{If a \code{NextPageToken} was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{nextPageToken}. Keep all other arguments unchanged.

The configured \code{maximumPageSize} determines how many results can be returned in a single call.}

\item{maximumPageSize}{The maximum number of results that are returned per call. \code{nextPageToken} can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size \emph{smaller} than the maximum.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.}

\item{reverseOrder}{When set to \code{true}, returns the results in reverse order. By default the results are returned in descending order of the start or the close time of the executions.}
}
\description{
Returns a list of closed workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.
}
\details{
This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

\strong{Access Control}

You can use IAM policies to control this action's access to Amazon SWF resources as follows:
\itemize{
\item Use a \code{Resource} element with the domain name to limit the action to only specified domains.
\item Use an \code{Action} element to allow or deny permission to call this action.
\item Constrain the following parameters by using a \code{Condition} element with the appropriate keys.
\item \code{tagFilter.tag}: String constraint. The key is \code{swf:tagFilter.tag}.
\item \code{typeFilter.name}: String constraint. The key is \code{swf:typeFilter.name}.
\item \code{typeFilter.version}: String constraint. The key is \code{swf:typeFilter.version}.
}

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's \code{cause} parameter is set to \code{OPERATION_NOT_PERMITTED}. For details and example IAM policies, see \href{http://docs.aws.amazon.com/amazonswf/latest/developerguide/swf-dev-iam.html}{Using IAM to Manage Access to Amazon SWF Workflows} in the \emph{Amazon SWF Developer Guide}.
}
\examples{

}
